Created inventory system alpha v0.01
Features: Press key I to open inventory. Click right mouse button on target item on the scene to quickly add item to the inventory. With the left button you can drag target item to the position in the inventory. After the inventory is closed, it is saved.
You can drop items from the inventory.
Works with icons and prefabs.